requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorators==2.0.7
deprecation==2.1.0
et_xmlfile==2.0.0
Faker==37.4.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
frozenlist==1.7.0
gotrue==2.12.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jupyterlab==4.3.3
MarkupSafe==3.0.2
multidict==6.5.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
notebook==7.3.1
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfkit==1.0.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==4.21.12
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.0
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
realtime==2.4.3
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.11.3
StrEnum==0.4.15
supafunc==0.9.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
webcolors==24.11.1
websockets==14.2
Werkzeug==3.1.3
XlsxWriter==3.2.3
yarl==1.20.1
xhtml2pdf==0.2.17